feat(api): add a basic ping endpoint
This adds the endpoint /api/v1/ping which will simply respond with
an `application/json` body of the form `{"pong": now}`, where `now`
is the current server time as returned by `Date.now()`. This can be
used for health or accidental caching checks as it is guaranteed to
be different as long as you wait at least 1ms between calls.