feat: store when each session token was last used
This is to allow for culling of unused session tokens. Later I plan
to add a cron-esque job that will run every day or so and delete any
session tokens where the last used date is older than a few months
(exact timings still tbd). This way active sessions won't time out on
the server side (though client cookies will still only persist if the
user chooses 'remember me' [implementation of that still in progress]),
but we have a way to keep the number of active rows in the database
to a manageable level.

It is worth noting that this does put more strain on the database, as
now every single authorized request will cause a database update. That
level of activity is fine for now, but it may need to be revisited in
the future.